requirement.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.1.1
django-sequences==2.6
sqlparse==0.4.3
